requirements.txt
# Memex dependencies
# Install all: pip install -r requirements.txt
# Or install what you need:

# MCP Server (required for Claude Code integration)
mcp>=0.9.0

# Hybrid Search (BM25 + semantic)
sentence-transformers>=2.2.0
# Or lighter/faster:
# fastembed>=0.2.0

# Batch API (50% cost reduction)
anthropic>=0.18.0
# Or:
# openai>=1.10.0

# PDF Ingestion
pymupdf>=1.24.0
# Optional for OCR:
# pytesseract
# pillow

# Voice Capture (local Whisper)
openai-whisper>=20231117
sounddevice>=0.4.6
soundfile>=0.12.0
numpy>=1.24.0
# Or for faster transcription:
# faster-whisper>=0.10.0

# Web Clipping
httpx>=0.26.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
# Or for better extraction:
# trafilatura>=1.6.0
